#28 fix: 로그인페이지 및 대회등록 페이지 폴더구조 변경 및 component화 완료
