+++ src/app/App.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import { Routes, Route} from "react-router-dom";
 import { UploadFileList } from "../features/post/upLoadFile/ui/upLoadFileList/UpLoadFileList";
 import { UpLoadButton } from "../features/post/upLoadFile/ui/UpLoadButton/UpLoadButton";
-import AddCompetitionPage from "../pages/competitionPages/ui/addCompetitionPage/AddCompetitionPage";
-import Footer from "../widgets/footer/ui/Footer"
-import LoginPage from "../pages/loginPage/ui/LoginPage";
+import {AddCompetitionPage} from "../pages/competitionPage";
+import {Footer} from "../widgets/footer"
+import {LoginPage} from "../pages/loginPage";
 import Main from "../pages/mainPage/ui/Main";
-import Header from "./header/Header";
-import {useAxiosInterceptor} from "./hocs/UseAxiosInterceptor";
+import {Header} from "../widgets/header";
+import {useAxiosInterceptor} from "../shared/api/UseAxiosInterceptor";
 import SignUpPage from "../pages/signUp/ui/SignUpPage/SignUpPage";
 
 function App() {
