--- src/app/appStore.ts
@@ -1,1 +1,0 @@
-export {}
