+++ src/features/ckEditor/index.ts
@@ -1,0 +1,1 @@
+export {CkEditor} from "./ui/CkEditor"
