+++ src/features/competition/index.ts
@@ -1,0 +1,2 @@
+export {AddFiles} from "./ui/AddFiles";
+export {PlaceArea} from "./ui/PlaceArea"
