+++ src/features/competition/ui/PlaceArea.module.css
