+++ src/pages/competitionPage/index.ts
@@ -1,0 +1,1 @@
+export {AddCompetitionPage} from "./ui/AddCompetitionPage"
