+++ src/pages/competitionPage/model/DivisionOptions.ts
@@ -1,0 +1,12 @@
+
+
+export const DivisionOptions: { value: string, label: string; }[] = [
+    {value: "element", label: "초등"},
+    {value: "middle", label: "중등"},
+    {value: "high", label: "고등"},
+    {value: "university", label: "대학"},
+    {value: "division3", label: "디비전3"},
+    {value: "division4", label: "디비전4"},
+    {value: "forties", label: "40대"},
+    {value: "rookie", label: "루키"},
+];
