+++ src/pages/competitionPage/ui/AddCompetitionPage.module.css
