+++ src/pages/loginPage/ui/LoginPage.tsx
@@ -1,11 +1,11 @@
 import React, {useEffect, useState} from 'react';
-import PageTitle from "../../../shared/ui/pageTitle/PageTitle";
+import {LoginInput, LoginLabel, PageTitle} from "../../../shared/ui";
 import style from "./LoginPage.module.css";
 import fetchLogin from "../api/FetchLogin";
-import useUserStore from "../../../app/hocs/UserStore";
+import {useUserStore} from "../../../shared/model";
 import {Link} from "react-router-dom";
 
-const LoginPage = () => {
+export const LoginPage = () => {
     const [email, setEmail] = useState<string>("");
     const [password, setPassword] = useState<string>("");
     const [emailMessage, setEmailMessage] = useState<string>("");
@@ -28,12 +28,8 @@ const LoginPage = () => {
             <PageTitle pageName={"로그인"}/>
             <form noValidate className={style.formBox} onSubmit={(e) => submitHandler(e)}>
                 <div className={style.boxArea}>
-                    <div className={style.labelArea}>
-                        <label>이메일</label>
-                    </div>
-                    <div className={style.inputArea}>
-                        <input type="email" onChange={(e) => setEmail(e.target.value)}/>
-                    </div>
+                    <LoginLabel name={"이메일"} />
+                    <LoginInput type={"email"} setFn={setEmail}/>
                 </div>
                 <div className={style.messageBox}>
                     <p className={style.message}>
@@ -42,12 +38,8 @@ const LoginPage = () => {
                 </div>
 
                 <div className={style.boxArea}>
-                    <div className={style.labelArea}>
-                        <label>비밀번호</label>
-                    </div>
-                    <div className={style.inputArea}>
-                        <input type="password" onChange={(e) => setPassword(e.target.value)}/>
-                    </div>
+                    <LoginLabel name={"비밀번호"} />
+                    <LoginInput type={"password"} setFn={setPassword}/>
                 </div>
                 <div className={style.searchBox}>
                     <Link to="/" className={style.search}>아이디 | 비밀번호 찾기 &gt;</Link>
@@ -59,5 +51,3 @@ const LoginPage = () => {
         </div>
     );
 };
-
-export default LoginPage;
