+++ src/shared/ui/index.ts
@@ -1,0 +1,4 @@
+export {AddCompetitionLabel} from "./addCompetitionLabel/AddCompetitionLabel"
+export {PageTitle} from "./pageTitle/PageTitle"
+export {LoginLabel} from "./loginLabel/LoginLabel"
+export {LoginInput} from "./loginInput/LoginInput"
