+++ src/shared/ui/pageTitle/PageTitle.tsx
@@ -4,12 +4,10 @@ import style from "./PageTitle.module.css"
 type Props = {
     pageName: string;
 }
-const PageTitle = ({pageName} : Props) => {
+export const PageTitle = ({pageName} : Props) => {
     return (
         <h1 className={style.page_title}>
             {pageName}
         </h1>
     );
 };
-
-export default PageTitle;
