+++ src/widgets/footer/index.ts
@@ -1,0 +1,1 @@
+export {Footer} from "./ui/Footer"
